+++ CHANGELOG.md
@@ -9,6 +9,9 @@
     if `ReferenceSubject` is null.
   - Fixed exception when Certificate with empty Subject exists in
     Certificate Store - fixes [Issue #190](https://github.com/PowerShell/CertificateDsc/issues/190).
+  - Fixed bugs matching existing certificate when Subject Alternate
+    Name is specified and machine language is not en-US - fixes
+    [Issue #193](https://github.com/PowerShell/CertificateDsc/issues/193).
 
 ## 4.5.0.0